+++ CHANGELOG.md
@@ -5,11 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added:
+* Utility to check if directory or file exists and create them
+
+### Changed:
+* Changed error handler logic
+* Updated types for error utility
+
 ## Patch 0.0.1-r
-## Added:
+### Added:
 * Error utils: isError, handle error sync and async, log error sync and async
+
 `08-05-2024`
 
 ## Patch 0.0.1
 **Changes: [(...)](https://github.com/stbestichhh/stlib-utils/commits/master/?since=2024-05-08&until=2024-05-08)**
+
 `06-05-2024`